+++ src/color/interfaces/IColorHSL.ts
@@ -4,4 +4,4 @@ export interface IColorHSLJson {
 	l: number;
 }
 
-export default interface IColorHSL extends Uint8Array, IColorHSLJson {}
+export default interface IColorHSL extends Float32Array, IColorHSLJson {}
